+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ### Changed
 - The `ResponseBuilder` now requires PSR-17 factories instead of a PHP-HTTP factory, to align it with the `Client`. This is only a breaking change in the rare case where you provide your own factory to the `ResponseBuilder`.
 - Removed `domainAliases` from constructor arguments. Please use `setDomainAliases()` on the instance instead.
+- Replaced [danielstjules/stringy](https://packagist.org/packages/danielstjules/stringy) with [symfony/string](https://packagist.org/packages/symfony/string). In some edge cases, mainly special or accented characters, this will affect how query parameters are transformed to a fixture.
 
 ### Removed
 - Dropped PHP <7.4 support.